+++ src/io/stdin.rs
@@ -1,7 +1,8 @@
 //! A handle to the standard input stream.
 use crate::{
-    alloc::NSTDAllocError,
+    alloc::{CBox, NSTDAllocError},
     core::{
+        optional::{gen_optional, NSTDOptional},
         result::NSTDResult,
         slice::{NSTDSlice, NSTDSliceMut},
         str::nstd_core_str_from_bytes_unchecked,
@@ -16,17 +17,26 @@ use std::io::{Stdin, StdinLock};
 use std::os::unix::io::AsRawFd;
 
 /// A handle to the standard input stream.
-pub type NSTDStdin = Box<Stdin>;
+#[nstdapi]
+pub struct NSTDStdin {
+    /// Rust's [Stdin].
+    r#in: CBox<Stdin>,
+}
+gen_optional!(NSTDOptionalStdin, NSTDStdin);
 
 /// Constructs a new handle to the standard input stream.
 ///
 /// # Returns
 ///
-/// `NSTDStdin handle` - A handle to the standard input stream.
+/// `NSTDOptionalStdin handle` - A handle to the standard input stream on success, or an
+/// uninitialized "none" variant on error.
 #[inline]
 #[nstdapi]
-pub fn nstd_io_stdin() -> NSTDStdin {
-    NSTDStdin::new(std::io::stdin())
+pub fn nstd_io_stdin() -> NSTDOptionalStdin {
+    match CBox::new(std::io::stdin()) {
+        Some(r#in) => NSTDOptional::Some(NSTDStdin { r#in }),
+        _ => NSTDOptional::None,
+    }
 }
 
 /// Reads some data from stdin into a byte slice buffer.
@@ -57,9 +67,9 @@ pub unsafe fn nstd_io_stdin_read(
     buffer: &mut NSTDSliceMut,
 ) -> NSTDIOResult {
     #[cfg(not(unix))]
-    return crate::io::stdio::read(handle, buffer);
+    return crate::io::stdio::read(&mut *handle.r#in, buffer);
     #[cfg(unix)]
-    return crate::os::unix::io::stdio::read(handle.lock().as_raw_fd(), buffer).into();
+    return crate::os::unix::io::stdio::read(handle.r#in.lock().as_raw_fd(), buffer).into();
 }
 
 /// Continuously reads data from stdin into a buffer until EOF is reached.
@@ -83,11 +93,11 @@ pub unsafe fn nstd_io_stdin_read(
 #[nstdapi]
 pub fn nstd_io_stdin_read_all(handle: &mut NSTDStdin, buffer: &mut NSTDVec) -> NSTDIOResult {
     #[cfg(not(unix))]
-    return crate::io::stdio::read_all(handle, buffer);
+    return crate::io::stdio::read_all(&mut *handle.r#in, buffer);
     #[cfg(unix)]
     // SAFETY: `handle` owns the file descriptor.
     unsafe {
-        crate::os::unix::io::stdio::read_all(handle.lock().as_raw_fd(), buffer).into()
+        crate::os::unix::io::stdio::read_all(handle.r#in.lock().as_raw_fd(), buffer).into()
     }
 }
 
@@ -115,11 +125,11 @@ pub fn nstd_io_stdin_read_to_string(
     buffer: &mut NSTDString,
 ) -> NSTDIOResult {
     #[cfg(not(unix))]
-    return crate::io::stdio::read_to_string(handle, buffer);
+    return crate::io::stdio::read_to_string(&mut *handle.r#in, buffer);
     #[cfg(unix)]
     // SAFETY: `handle` owns the file descriptor.
     unsafe {
-        crate::os::unix::io::stdio::read_to_string(handle.lock().as_raw_fd(), buffer).into()
+        crate::os::unix::io::stdio::read_to_string(handle.r#in.lock().as_raw_fd(), buffer).into()
     }
 }
 
@@ -150,9 +160,9 @@ pub unsafe fn nstd_io_stdin_read_exact(
     buffer: &mut NSTDSliceMut,
 ) -> NSTDIOError {
     #[cfg(not(unix))]
-    return crate::io::stdio::read_exact(handle, buffer);
+    return crate::io::stdio::read_exact(&mut *handle.r#in, buffer);
     #[cfg(unix)]
-    return crate::os::unix::io::stdio::read_exact(handle.lock().as_raw_fd(), buffer).into();
+    return crate::os::unix::io::stdio::read_exact(handle.r#in.lock().as_raw_fd(), buffer).into();
 }
 
 /// Reads a line from stdin and appends it to `buffer`.
@@ -170,7 +180,7 @@ pub unsafe fn nstd_io_stdin_read_exact(
 #[nstdapi]
 pub fn nstd_io_stdin_read_line(handle: &mut NSTDStdin, buffer: &mut NSTDString) -> NSTDIOResult {
     let mut buf = String::new();
-    match handle.read_line(&mut buf) {
+    match handle.r#in.read_line(&mut buf) {
         Ok(r) => {
             let bytes = NSTDSlice::from_slice(buf.as_bytes());
             // SAFETY: `bytes` refers to `buf`'s data, which is still valid UTF-8 here.
@@ -197,17 +207,26 @@ pub fn nstd_io_stdin_read_line(handle: &mut NSTDStdin, buffer: &mut NSTDString) 
 pub fn nstd_io_stdin_free(handle: NSTDStdin) {}
 
 /// A locked handle to the standard input stream.
-pub type NSTDStdinLock = Box<StdinLock<'static>>;
+#[nstdapi]
+pub struct NSTDStdinLock {
+    /// Rust's [StdinLock].
+    r#in: CBox<StdinLock<'static>>,
+}
+gen_optional!(NSTDOptionalStdinLock, NSTDStdinLock);
 
 /// Constructs a new locked handle to the standard input stream.
 ///
 /// # Returns
 ///
-/// `NSTDStdinLock handle` - A locked handle to the standard input stream.
+/// `NSTDOptionalStdinLock handle` - A locked handle to the standard input stream on success, or an
+/// uninitialized "none" variant on error.
 #[inline]
 #[nstdapi]
-pub fn nstd_io_stdin_lock() -> NSTDStdinLock {
-    NSTDStdinLock::new(std::io::stdin().lock())
+pub fn nstd_io_stdin_lock() -> NSTDOptionalStdinLock {
+    match CBox::new(std::io::stdin().lock()) {
+        Some(r#in) => NSTDOptional::Some(NSTDStdinLock { r#in }),
+        _ => NSTDOptional::None,
+    }
 }
 
 /// Reads some data from stdin into a byte slice buffer.
@@ -238,9 +257,9 @@ pub unsafe fn nstd_io_stdin_lock_read(
     buffer: &mut NSTDSliceMut,
 ) -> NSTDIOResult {
     #[cfg(not(unix))]
-    return crate::io::stdio::read(handle, buffer);
+    return crate::io::stdio::read(&mut *handle.r#in, buffer);
     #[cfg(unix)]
-    return crate::os::unix::io::stdio::read(handle.as_raw_fd(), buffer).into();
+    return crate::os::unix::io::stdio::read(handle.r#in.as_raw_fd(), buffer).into();
 }
 
 /// Continuously reads data from stdin into a buffer until EOF is reached.
@@ -267,11 +286,11 @@ pub fn nstd_io_stdin_lock_read_all(
     buffer: &mut NSTDVec,
 ) -> NSTDIOResult {
     #[cfg(not(unix))]
-    return crate::io::stdio::read_all(handle, buffer);
+    return crate::io::stdio::read_all(&mut *handle.r#in, buffer);
     #[cfg(unix)]
     // SAFETY: `handle` owns the file descriptor.
     unsafe {
-        crate::os::unix::io::stdio::read_all(handle.as_raw_fd(), buffer).into()
+        crate::os::unix::io::stdio::read_all(handle.r#in.as_raw_fd(), buffer).into()
     }
 }
 
@@ -299,11 +318,11 @@ pub fn nstd_io_stdin_lock_read_to_string(
     buffer: &mut NSTDString,
 ) -> NSTDIOResult {
     #[cfg(not(unix))]
-    return crate::io::stdio::read_to_string(handle, buffer);
+    return crate::io::stdio::read_to_string(&mut *handle.r#in, buffer);
     #[cfg(unix)]
     // SAFETY: `handle` owns the file descriptor.
     unsafe {
-        crate::os::unix::io::stdio::read_to_string(handle.as_raw_fd(), buffer).into()
+        crate::os::unix::io::stdio::read_to_string(handle.r#in.as_raw_fd(), buffer).into()
     }
 }
 
@@ -334,9 +353,9 @@ pub unsafe fn nstd_io_stdin_lock_read_exact(
     buffer: &mut NSTDSliceMut,
 ) -> NSTDIOError {
     #[cfg(not(unix))]
-    return crate::io::stdio::read_exact(handle, buffer);
+    return crate::io::stdio::read_exact(&mut *handle.r#in, buffer);
     #[cfg(unix)]
-    return crate::os::unix::io::stdio::read_exact(handle.as_raw_fd(), buffer).into();
+    return crate::os::unix::io::stdio::read_exact(handle.r#in.as_raw_fd(), buffer).into();
 }
 
 /// Frees and unlocks an instance of `NSTDStdinLock`.