+++ CHANGELOG.md
@@ -1,7 +1,36 @@
 # TBD
+## Added
+### `nstd`
+- Added an `nstd` feature to enable all modules.
+### `nstd.io`
+- Added `NSTDOptionalStd[in|out|err][Lock]`.
+### `nstd.mutex`
+- Added `NSTDOptionalMutex`.
+### `nstd.thread`
+- Added `NSTDOptionalThreadHandle`.
+- Added `NSTDOptionalThreadID`.
 ## Changed
+### `nstd`
+- Removed the `nstd_` prefix from all features.
 ### `nstd.env`
 - `nstd_env_temp_dir` now returns `NSTDOptionalString`.
+### `nstd.io`
+- `nstd_io_std[in|out|err]` functions now return `NSTDOptionalStd[in|out|err]` respectively.
+- `nstd_io_std[in|out|err]_lock` functions now return `NSTDOptionalStd[in|out|err]Lock`
+respectively.
+### `nstd.mutex`
+- `nstd_mutex_new` now returns `NSTDOptionalMutex`.
+- `nstd_mutex_lock` now returns `NSTDOptionalMutexLockResult`.
+### `nstd.thread`
+- `NSTDThreadResult` is now `NSTDOptionalHeapPtr`.
+- `nstd_thread_spawn` now takes `const NSTDThreadDescriptor *desc`.
+- `nstd_thread_[current|handle]` functions now return `NSTDOptionalThreadHandle`.
+- `nstd_thread_id` now returns `NSTDOptionalThreadID`.
+## Removed
+### `nstd`
+- Removed the `asm` feature.
+### `nstd.thread`
+- `nstd_thread_spawn_with_desc`.
 
 # 0.8.0
 ## Added